// frontend/ticket-booking-web/src/app/api/model/add-ticket-category-request.ts
/**
 * Ticket Booking Platform API - v1
 *
 * Contact: m1musbal@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AddTicketCategoryRequest { 
    name: string;
    price: number;
    currency: string;
    quantity: number;
}