// frontend/ticket-booking-web/src/app/api/model/ticket-category-dto.ts
/**
 * Ticket Booking Platform API - v1
 *
 * Contact: m1musbal@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TicketCategoryDto { 
    id: string;
    name: string;
    price: number;
    currency: string;
    totalQuantity: number;
    availableQuantity: number;
}